+++ playground/Coalesce.Web.Vue3/Coalesce.Web.Vue3.csproj
@@ -7,7 +7,12 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>true</ImplicitUsings>
   </PropertyGroup>
-
+  
+  <ItemGroup>
+    <!--<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.2" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="6.0.3" />
+    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="6.0.1" />-->
+  </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Coalesce.Domain\Coalesce.Domain.csproj" />
     <ProjectReference Include="..\..\src\IntelliTect.Coalesce.Vue\IntelliTect.Coalesce.Vue.csproj" />
