+++ playground/Coalesce.Web.Vue3/Program.cs
@@ -1,7 +1,7 @@
 using Coalesce.Domain;
 using Coalesce.Domain.WebShared;
 using IntelliTect.Coalesce;
-using Microsoft.OpenApi;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging.Console;
@@ -11,8 +11,6 @@ using System.Security.Claims;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
-using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 
 var builder = WebApplication.CreateBuilder(new WebApplicationOptions
 {
