+++ src/IntelliTect.Coalesce/Models/ItemResult.cs
@@ -231,7 +231,12 @@ namespace IntelliTect.Coalesce.Models
             foreach (var param in method.ClientParameters.OfType<ReflectionParameterViewModel>())
             {
                 string propName = param.Name;
-                var value = obj.GetType().GetProperty(param.PascalCaseName)?.GetValue(obj);
+                Type type = obj.GetType();
+
+                // Checking CsParameterName is backwards-compat for when we generated parms object with camelCase names.
+                var propertyInfo = type.GetProperty(param.PascalCaseName) ?? type.GetProperty(param.CsParameterName);
+
+                var value = propertyInfo?.GetValue(obj);
 
                 var validationContext = new ValidationContext(obj, serviceProvider, null)
                 {
